refactor(i18n): restructure language management and context initialization - Removed the setLanguage function from RottUiContextModel. - Introduced initialState for RottUiContext to streamline context creation. - Moved languageMessages import to constants and updated related imports. - Deleted unused I18nProvider and its associated files to simplify the i18n structure.
