+++ src/libs/i18n/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './languageMessages'
