+++ src/libs/i18n/constants/languageMessages.ts
@@ -1,0 +1,7 @@
+import en from '../en-US.json'
+import tr from '../tr-TR.json'
+
+export const languageMessages = {
+  'tr-TR': tr,
+  'en-US': en,
+}
