+++ src/libs/i18n/index.ts
@@ -1,4 +1,4 @@
 export * from './models'
-export * from './providers'
 export * from './utils'
 export * from './hooks'
+export * from './constants'
