--- src/libs/i18n/providers/index.ts
@@ -1,1 +1,0 @@
-export * from './I18Provider'
